accountant

+++ database/migrations/2024_09_18_113729_create_orders_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
                 ->onDelete('cascade');
             $table->unsignedBigInteger('doctor_id');
             $table->unsignedBigInteger('type_id');
-            $table->integer('paid')->default(0);
+            $table->decimal('paid')->default(0);
             $table->boolean('invoiced')->default(false);
             $table->decimal('cost',15,2);
             $table->string('patient_name');
